Put "long()" around default value for "seed"...

+++ Enumerate.py
@@ -175,10 +175,10 @@ if __name__ == "__main__":
   # Get command line options.
   # Enumerate.py may be sent an infile param but it should be ignored
   opts, args = Generator_Option_Parser(sys.argv, 
-          [("time",       int,           15, False, True), 
-           ("save_freq",  int,       100000, False, True),
-           ("seed",      long,  time.time(), False, True),
-           ("checkpoint", str, "checkpoint", False, True)],
+          [("time",       int,                15, False, True), 
+           ("save_freq",  int,            100000, False, True),
+           ("seed",      long, long(time.time()), False, True),
+           ("checkpoint", str,      "checkpoint", False, True)],
           ignore_infile=True)
   
   steps = (opts["steps"] if opts["steps"] > 0 else Macro_Simulator.INF)
